Add FromPrimitive and ToPrimitive to enum derives and use enums directly as function parameters

+++ rust-package/brainflow/examples/markers.rs
@@ -8,7 +8,7 @@ use brainflow::BoardIds;
 fn main() {
     brainflow::board_shim::enable_dev_board_logger().unwrap();
     let params = BrainFlowInputParamsBuilder::default().build();
-    let board = BoardShim::new(BoardIds::SyntheticBoard as i32, params).unwrap();
+    let board = BoardShim::new(BoardIds::SyntheticBoard, params).unwrap();
 
     board.prepare_session().unwrap();
     board.start_stream(45000, "file://data.csv:w").unwrap();
